webant test base class
some tests were working on the default `libreant` index and they did not delete it after they've finished.

Now the WebantTestCase class make use of a temporary index that will be deeleted at the end of the class level tests.

Moreover now it's easier to subclass the WebantTestCase class.
As you can see in the `WebantUsersTestCase` class it is possible to use the `setUpClass` method to override the webant configuration that are going to be use during the tests.